Numeric Matrix Processor/Matrix by matrix multiplication(stage 3/6): input_matrix() method updated
- accepts floating point numbers when filling a matrix

+++ Numeric Matrix Processor/task/processor/processor.py
@@ -181,7 +181,7 @@ def input_matrix(part_of_message):
     print(f"Enter {part_of_message}: ")
 
     for i in range(1, rows + 1):
-        row = [int(item) for item in input().split()]
+        row = [float(item) for item in input().split()]
         res.set_row(i, row)
 
     return res
